+++ backend/src/user/user.gateway.ts
@@ -1,8 +1,7 @@
 import { Server } from 'socket.io';
 import { WebSocketGateway, WebSocketServer } from '@nestjs/websockets';
 
-import { SocketId } from '../util/type';
-import { UserId } from '../util/type';
+import { SocketId, UserId } from '../util/type';
 import { UserInfoDto } from './dto/user-gateway.dto';
 
 @WebSocketGateway()
