+++ backend/src/user/user.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 
 import { UserGateway } from './user.gateway';
-import { UserStatusModule } from '../user-status/user-status.module';
 import { UserService } from './user.service';
+import { UserStatusModule } from '../user-status/user-status.module';
 
 @Module({
   imports: [UserStatusModule],
